Adjusted the input data type for fixed size programs
The fixed size parameter for programs should be unsigned rather than
signed. This allows to pass size_t parameters directly.

+++ include/beast/program_factory_base.hpp
@@ -23,7 +23,7 @@ class ProgramFactoryBase {
    * @param size tbd
    * @return tbd
    */
-  virtual Program generate(size_t size) = 0;
+  virtual Program generate(uint32_t size) = 0;
 };
 
 }  // namespace beast
